requirements.txt
fastapi
uvicorn[standard]
authlib
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic